# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Platform API
  description: |
    Phase 1 backend API — auth, restaurant leads, durable city scraping, OCR/review gates,
    demo sites, and **email outreach campaigns**.

    This document covers the Phase 1 lead-to-outreach and reservation surface. The router
    remains the source of truth for separate company-consultation and legacy public asset
    routes that are outside this workflow specification.

    ## Authentication
    Protected routes use `Authorization: Bearer <access_token>` from signup or login.

    ## Roles
    - `internal_admin` — full restaurant admin (seed via `make seed-admin`)
    - `restaurant_owner` — scoped to assigned restaurants only
    - `developer` — health endpoints (local/test signup only)

    ## Lead preparation and outreach flow (admin)
    1. `POST /api/v1/scrape-jobs` — create or return the active durable city/niche scrape job.
    2. The scrape worker uses Google Places first and Apollo only for missing contact details.
       It pauses at 500 combined provider requests and resumes after its durable `resume_at`.
    3. OCR moves each profile through `pending`, `running`, then `verified`, `no_images`, or
       `failed`. Only `verified` can advance to outreach.
    4. OCR verification automatically creates or refreshes draft demo/campaign artifacts.
    5. Inspect the protected profile and demo review-preview endpoints; neither discloses
       the demo bearer token.
...

...

...

    ## Local email testing
...
servers:
...
paths:
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Sign up
      description: |
        Self-signup for `restaurant_owner` (all envs) and `developer` (local/test only).
        `internal_admin` cannot be self-assigned — use `make seed-admin`.
      operationId: authSignup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              email: owner@local.test
              password: password123
              full_name: Demo Restaurant Owner
              role: restaurant_owner
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@local.test
              password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current session identity
      operationId: authMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Principal from JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/me:
    get:
      tags: [Admin]
      summary: Admin profile
      operationId: adminMe
      security:
        - bearerAuth: []
      description: Requires role `internal_admin`.
      responses:
        '200':
          description: Admin user record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/user/me:
    get:
      tags: [User]
      summary: User profile
      operationId: userMe
      security:
        - bearerAuth: []
      description: |
        Requires role `restaurant_owner` or `developer`.
        Returns full user account details and restaurants where the user is a member.
        For JWT claims only, use `GET /api/v1/auth/me`. Admins use `GET /api/v1/admin/me`.
      responses:
        '200':
          description: User profile with restaurant memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                email: owner@local.test
                full_name: Demo Restaurant Owner
                role: restaurant_owner
                is_active: true
                created_at: '2026-06-24T10:00:00Z'
                updated_at: '2026-06-24T10:00:00Z'
                restaurants:
                  - id: 660e8400-e29b-41d4-a716-446655440001
                    name: Demo Fixture Cafe
                    email: owner@local.test
                    status: lead
                    member_role: owner
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/restaurants:
    get:
      tags: [Restaurants]
      summary: List accessible restaurants
      operationId: listRestaurants
      security:
        - bearerAuth: []
      description: |
        - `internal_admin` — all restaurants (subject to filters)
        - `restaurant_owner` — only restaurants where user is a member
        Archived restaurants are hidden unless `include_archived=true`.
      parameters:
        - $ref: '#/components/parameters/restaurantNameFilter'
        - $ref: '#/components/parameters/statusFilter'
        - $ref: '#/components/parameters/isContactedFilter'
        - $ref: '#/components/parameters/shownInterestFilter'
        - $ref: '#/components/parameters/ocrStatusFilter'
        - $ref: '#/components/parameters/includeArchivedFilter'
      responses:
        '200':
          description: Restaurant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Restaurants]
      summary: Create restaurant lead
      operationId: createRestaurant
      security:
        - bearerAuth: []
      description: Requires role `internal_admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
            example:
              name: PM Demo Thai Kitchen
              email: lead@pmdemo.test
      responses:
        '201':
          description: Restaurant created (status defaults to `lead`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/restaurants/{id}:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    get:
      tags: [Restaurants]
      summary: Get restaurant by ID
      operationId: getRestaurant
      security:
        - bearerAuth: []
      description: |
        Admin or restaurant owner with membership.
        Archived restaurants return 404 for non-admin users.
      responses:
        '200':
          description: Restaurant record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Restaurants]
      summary: Update restaurant fields
      operationId: updateRestaurant
      security:
        - bearerAuth: []
      description: |
        Requires role `internal_admin`. The admin UI treats `is_contacted` and
        `shown_interest` as read-only automation results: confirmed provider sends set
        contacted/emailed, while tracked link clicks set shown-interest/interested.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRestaurantRequest'
            example:
              is_contacted: true
      responses:
        '200':
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Restaurants]
      summary: Archive restaurant (soft delete)
      operationId: archiveRestaurant
      security:
        - bearerAuth: []
      description: Sets `status` to `archived`. Requires role `internal_admin`.
      responses:
        '200':
          description: Archived restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/restaurants/{id}/status:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    patch:
      tags: [Restaurants]
      summary: Update restaurant lifecycle status
      operationId: updateRestaurantStatus
      security:
        - bearerAuth: []
      description: Requires role `internal_admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRestaurantStatusRequest'
            example:
              status: demo_ready
      responses:
        '200':
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/restaurants/{id}/members:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    get:
      tags: [Members]
      summary: List restaurant members
      operationId: listRestaurantMembers
      security:
        - bearerAuth: []
      description: Requires role `internal_admin`.
      responses:
        '200':
          description: Membership list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Members]
      summary: Add restaurant member
      operationId: addRestaurantMember
      security:
        - bearerAuth: []
      description: |
        Requires role `internal_admin`.
        Links a user (e.g. restaurant owner) to a restaurant for tenant scoping.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
            example:
              user_id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              member_role: owner
      responses:
        '201':
          description: Member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/restaurants/{id}/demo-sites:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    post:
      tags: [Demo Sites]
      summary: Create demo site
      operationId: createDemoSite
      security:
        - bearerAuth: []
      description: |
        Requires role `internal_admin`.
        Creates a **draft-only** demo and returns a one-time `token` in the response.
        If `public_payload` is omitted, the server snapshots the restaurant's current
        public-safe profile, menu, image, and review data. It does not create a sample restaurant.

        Supplying `status=published` is rejected. Publish through
        `PATCH /api/v1/demo-sites/{id}/status` so OCR verification, human profile
        approval, and the publishing actor/timestamp are enforced atomically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDemoSiteRequest'
            example:
              slug: pm-demo-thai-kitchen
              status: draft
      responses:
        '201':
          description: Demo site created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDemoSiteResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/restaurants/{id}/profile/review-preview:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    get:
      tags: [Lead Review]
      summary: Inspect a profile and its review versions
      operationId: getRestaurantProfileReviewPreview
      security:
        - bearerAuth: []
      description: |
        Requires role `internal_admin`. Returns the reviewed restaurant name/contact,
        OCR/review state, allowlisted profile fields, and both version timestamps needed
        for an attributable decision. No raw Apollo payload or demo bearer token is returned.
        The response is marked `Cache-Control: no-store`.
      responses:
        '200':
          description: Review-safe restaurant/profile artifact
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always `no-store`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileReviewPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from signup or login
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SignupRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/SignupRole'
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...